pyproject.toml
[tool.poetry]
name = "faebot"
version = "0.1.0"
description = "discord bot for playing slightly modified Fate Accelerated Edition"
authors = ["SilentWrangler <thousandgoblins@gmail.com>"]
license = "CC BY 4.0"

[tool.poetry.dependencies]
python = ">=3.8.6,<4.0"
SQLAlchemy = "^1.4.26"
alembic = "^1.7.4"
contextlib3 = "^3.10.0"
python-dotenv = "^0.19.1"
#mysqlclient = "^2.0.3"
discord-py-interactions = "^4.3.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
